requirements.txt
appwrite
python-telegram-bot
requests